+++ apps/api/nestjs/src/app.module.ts
@@ -4,9 +4,10 @@ import { AppService } from './app.service.js';
 import { AuthModule } from './auth/auth.module.js';
 import { CatalogModule } from './catalog/catalog.module.js';
 import { PrismaModule } from './prisma/prisma.module.js';
+import { StockModule } from './stock/stock.module.js';
 
 @Module({
-  imports: [PrismaModule, AuthModule, CatalogModule],
+  imports: [PrismaModule, AuthModule, CatalogModule, StockModule],
   controllers: [AppController],
   providers: [AppService],
 })